Staging

+++ app/Http/Controllers/AdminSettingsController.php
@@ -290,7 +290,9 @@ class AdminSettingsController extends Controller
             $files = $input['default_video_image'];
             $format=$files->getClientOriginalExtension();
             $filename ='default_image'.'.' . 'jpg';
-            unlink( public_path().'/uploads/images/'.$defaultImage_setting); // Remove Image
+            if(!file_exists(public_path().'/uploads/images/'.$defaultImage_setting)){
+                unlink( public_path().'/uploads/images/'.$defaultImage_setting); // Remove Image
+            }
             Image::make($files)->save(base_path().'/public/uploads/images/'.$filename )->encode('jpg', 80);
             $settings->default_video_image = $filename;
       }
